Move GUI code as a submodule of params
